feat(campaign-users) add Many-to-Many relationship between Users and Campaigns
The User and Campaign models have been modified to include a reference to each other using GORM Many-to-Many relationship model. This will allow managing the user-campaign relations more efficiently.
In the database initialization script, an auto-migration function for an associative entity CampaignsUsers, which is designed to hold the relationship details between the User and Campaign models, has also been included.