feat(communities): add user registration to community feature
This commit introduces a feature where a user can register to a community. A new function `RegisterUserToCommunity` was added to `communityService.go` which saves a record of a User ID against a Community ID in the database. This function is called in a new route `/:id/register-user`, added to `communityRoutes.go`, that posts the registration using `RegisterUserToCommunity` handler defined in `communityHandler.go`. This ensures a user can join a community.