refactor(core-client): Refactor codes formats and apply linter configurations
Reformatted the code to provide consistency in its format such as the use of double-quote for strings and removing the semicolon at the end of each expression to satisfy the Prettier and ESLint rules. These changes enhance code readability and maintainability. Also added new devDependencies that include linter configurations. This will ensure that future modification will compliant with the defined code standard rules.