feat(users_service.go): add UserIsInCommuinty function to check if a user is in a community
The UserIsInCommuinty function is added to the users_service.go file. This function takes a user and a community ID as parameters and checks if the user is a member of the community. It iterates through the user's communities and returns true if a match is found, otherwise it returns false. This function is useful for checking if a user has access to certain community-specific features or resources.