feat(community_middleware.go): add community middleware to check if user is in the community before allowing access to certain routes
The `CommunityMiddleware` function is added to the `middleware` package. This middleware function checks if the user is authenticated and if they are a member of the community specified by the `id` parameter. If the user is not authenticated or not a member of the community, a `401 Unauthorized` status is returned. Otherwise, the middleware calls `c.Next()` to allow the request to proceed to the next handler.